+++ internal/v1/models/product.go
