+++ internal/v1/product/routes.go
